Twig functions: separate from Pixelssite to own class
